adjust the header hight
